module1_scheme_advisor/requirement.txt
# Module 3: Government Scheme Advisor - Dependencies

# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.13

# Vector Database
faiss-cpu==1.7.4

# Embeddings
sentence-transformers==2.2.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# LLM Integration (Free!)
groq==0.4.2

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
tiktoken==0.5.2

# Optional: For better PDF extraction
pypdf==3.17.4
